+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Not released
 
+- Change `.first` to not query the API if records have already been retrieved (https://github.com/Beyond-Finance/active_force/pull/73)
+
 ## 0.19.0
 
 - Bulk API methods. (https://github.com/Beyond-Finance/active_force/pull/65)